upd multisig examples: misc upd
